finally got these changes to run
problem here was that groupby_random_choice doesn't support choosing from a group that doesn't exist.  in the proportional model we sometimes ask to increase retail or gov/edu in a city / zone_id which does not have the correct building types

this also requires changes in orca that are still being litigated

+++ baus/variables.py
@@ -110,7 +110,7 @@ def job_spaces(buildings):
 @orca.column('buildings')
 def vacant_job_spaces(buildings, jobs):
     return buildings.job_spaces.sub(jobs.building_id.value_counts(),
-                                    fill_value=0)
+                                    fill_value=0).astype('int')
 
 
 @orca.column('buildings', cache=True)
